+++ lib/api/src/lib.rs
@@ -1,4 +1,8 @@
 //! Wasmer API
+#![doc(
+    html_logo_url = "https://github.com/wasmerio.png?size=200",
+    html_favicon_url = "https://wasmer.io/static/icons/favicon.ico"
+)]
 #![deny(
     missing_docs,
     trivial_numeric_casts,